Add CI, add npmignore, make better testing
